requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
pyserial==3.5
python-dotenv==1.0.0
